app/requirements.txt
fastapi
uvicorn
transformers
Pillow
pydantic
pyyaml
huggingface_hub